Fonts, levels, fixes

+++ TwoButton/Assets/Mad Level Manager/Scripts/MadCommons/Editor/MadGUI.cs
@@ -80,6 +80,7 @@ public class MadGUI {
     
         public string label = "Scrollable List";
         public int height = 200;
+        public int spaceAfter = 0;
         
         public bool selectionEnabled = false;
         public string emptyListMessage = "No elements!";
@@ -116,69 +117,76 @@ public class MadGUI {
         public ScrollableList(List<T> items) {
             this.items = items;
         }
-        
+
         public void Draw() {
             if (skin == null) {
                 skin = Resources.Load("GUISkins/editorSkin", typeof(GUISkin)) as GUISkin;
             }
 
             var allRect = EditorGUILayout.BeginVertical();
-        
+
             bool toggleColor = false;
             Color baseColor = GUI.color;
-        
+
             GUILayout.Label(label);
             position = EditorGUILayout.BeginScrollView(
                 position, false, true, GUILayout.Height(height));
-                
+
             float y = 0;
-                
+
             foreach (var item in items) {
-                var rect = EditorGUILayout.BeginHorizontal();
+                var itemRect = EditorGUILayout.BeginHorizontal();
                 if (scrollToItem == item && Event.current.type == EventType.Repaint) {
-                    if (y + rect.height > allRect.height || y < 0) {
+                    if (y - position.y < 0) { // item above
                         position.y = y;
+                    } else if (y - position.y + itemRect.height + 2 >= allRect.height) { // item below ( + 2 - hack)
+                        position.y = y - (allRect.height - itemRect.height * 2);
                     }
+
                     scrollToItem = null;
                 }
 
                 if (selectionEnabled) {
                     if (Event.current.type == EventType.MouseDown && Event.current.button == 0) {
-                        if (rect.Contains(Event.current.mousePosition)) {
+                        if (itemRect.Contains(Event.current.mousePosition)) {
                             selectedItem = item;
                         }
                     }
                 }
-                
+
                 // component value based on skin
                 float c = EditorGUIUtility.isProSkin ? 0 : 1;
-                
+
                 GUI.color = toggleColor ? Color.clear : new Color(c, c, c, 0.2f);
                 toggleColor = !toggleColor;
                 if (item.selected) {
                     GUI.color = toggleColor ? new Color(c, c, c, 0.4f) : new Color(c, c, c, 0.6f);
                 }
-                
-                GUI.Box(rect, "", skin.box);
+
+                GUI.Box(itemRect, "", skin.box);
                 GUI.color = baseColor;
-                
+
                 EditorGUILayout.BeginVertical();
                 item.OnGUI();
                 EditorGUILayout.EndVertical();
 
                 EditorGUILayout.EndHorizontal();
-                
-                y += rect.height;
+
+                y += itemRect.height;
             }
-            
+
             if (items.Count == 0) {
                 GUILayout.Label(emptyListMessage);
             }
 
             GUILayout.FlexibleSpace();
-                
+
+            if (spaceAfter != 0) {
+                GUILayout.Space(spaceAfter);
+            }
+
             EditorGUILayout.EndScrollView();
-            
+
             GUI.color = baseColor;
 
             EditorGUILayout.EndVertical();
@@ -246,6 +254,7 @@ public class MadGUI {
 		public string addLabel = "Add";
 
         public bool drawSeparator = false;
+        public bool drawOrderButtons = true;
         
         public RunnableGeneric0<T> createFunctionGeneric = () => { return new T(); };
         public RunnableVoid1<SerializedProperty> createFunctionProperty = (element) => {
@@ -261,11 +270,11 @@ public class MadGUI {
             }
         };
         
-        public RunnableVoid0 beforeAdd = () => {};
-        public RunnableVoid1<T> beforeRemove = (arg1) => {};
+        public RunnableVoid0 beforeAdd;
+        public RunnableVoid1<T> beforeRemove;
         
-        public RunnableVoid1<T> onAdd = (arg1) => {};
-        public RunnableVoid1<T> onRemove = (arg1) => {};
+        public RunnableVoid1<T> onAdd;
+        public RunnableVoid1<T> onRemove;
         
         public ArrayList(SerializedProperty arrayProperty, RunnableVoid1<SerializedProperty> renderer) {
             this.arrayProperty = arrayProperty;
@@ -303,7 +312,9 @@ public class MadGUI {
         }
         
         void AddItem() {
-            beforeAdd();
+            if (beforeAdd != null) {
+                beforeAdd();
+            }
         
             if (items != null) {
                 items.Add(createFunctionGeneric());
@@ -313,8 +324,10 @@ public class MadGUI {
                 createFunctionProperty(prop);
             }
 
-            var item = ItemAt(ItemCount() - 1);
-            onAdd(item);
+            if (onAdd != null) {
+                var item = ItemAt(ItemCount() - 1);
+                onAdd(item);
+            }
         }
         
         void RenderItem(int index) {
@@ -341,7 +354,34 @@ public class MadGUI {
                     EditorGUILayout.BeginVertical();
                     RenderItem(i);
                     EditorGUILayout.EndVertical();
-                    
+
+                    GUI.backgroundColor = Color.yellow;
+
+                    if (drawOrderButtons) {
+
+                        string upLabel;
+                        string downLabel;
+
+#if !UNITY_3_5
+                    upLabel = "\u25B2";
+                    downLabel = "\u25BC";
+#else
+                        upLabel = "U";
+                        downLabel = "D";
+#endif
+                        GUI.enabled = CanMoveDown(i);
+                        if (GUILayout.Button(downLabel, GUILayout.ExpandWidth(false))) {
+                            MoveDown(i);
+                        }
+
+                        GUI.enabled = CanMoveUp(i);
+                        if (GUILayout.Button(upLabel, GUILayout.ExpandWidth(false))) {
+                            MoveUp(i);
+                        }
+
+                        GUI.enabled = true;
+                    }
+
                     GUI.backgroundColor = Color.red;
                     if (GUILayout.Button("X", GUILayout.ExpandWidth(false))) {
                         removeIndex = i;
@@ -359,10 +399,20 @@ public class MadGUI {
                 }
                 
                 if (removeIndex != -1) {
-                    T item = ItemAt(removeIndex);
-                    beforeRemove(item);
+                    T item = null;
+                    if (beforeRemove != null || beforeAdd != null) {
+                        item = ItemAt(removeIndex);
+                    }
+
+                    if (beforeRemove != null) {
+                        beforeRemove(item);
+                    }
+
                     RemoveItem(removeIndex);
-                    onRemove(item);
+
+                    if (onRemove != null) {
+                        onRemove(item);
+                    }
                 }
             }
 
@@ -376,6 +426,35 @@ public class MadGUI {
             EditorGUILayout.EndHorizontal();
             
             return EditorGUI.EndChangeCheck();
+        }
+
+        private bool CanMoveUp(int i) {
+            return i > 0;
+        }
+
+        private bool CanMoveDown(int i) {
+            return i < ItemCount() - 1;
+        }
+
+        private void MoveUp(int i) {
+            if (items != null) {
+                T tmp = items[i];
+                items[i] = items[i - 1];
+                items[i - 1] = tmp;
+            } else {
+                arrayProperty.MoveArrayElement(i, i - 1);
+            }
+            
+        }
+
+        private void MoveDown(int i) {
+            if (items != null) {
+                T tmp = items[i];
+                items[i] = items[i + 1];
+                items[i + 1] = tmp;
+            } else {
+                arrayProperty.MoveArrayElement(i, i + 1);                
+            }
         }
     }
 
@@ -431,6 +510,10 @@ public class MadGUI {
 
     public static void IndentBox(RunnableVoid0 runnable) {
         IndentBox("", runnable);
+    }
+
+    public static IndentC Indent() {
+        return new IndentC();
     }
     
     public static void IndentBox(string label, RunnableVoid0 runnable) {
@@ -619,8 +702,8 @@ public class MadGUI {
         }
         EditorGUILayout.EndHorizontal();
     }
-    
-    public static void PropertyFieldEnumPopup(SerializedProperty obj, string label) {
+
+    public static void PropertyFieldEnumPopup(SerializedProperty obj, string label, params GUILayoutOption[] options) {
         var names = obj.enumNames;
         // split names by camel-case
         for (int i = 0; i < names.Length; ++i) {
@@ -631,7 +714,11 @@ public class MadGUI {
                 bool first = j == 0;
                 char p = first ? ' ' : name[j - 1];
                 char c = name[j];
-                if (!first && char.IsUpper(c) && !char.IsUpper(p)) {
+                if (
+                    (!first && char.IsUpper(c) && !char.IsUpper(p)) // upper case check
+                    ||
+                    (!first && char.IsNumber(c) && !char.IsNumber(p)) // number check
+                    ) {
                     newName.Append(" ");
                 }
                 
@@ -642,7 +729,7 @@ public class MadGUI {
         }
         
         int selectedIndex = obj.enumValueIndex;
-        int newIndex = EditorGUILayout.Popup(label, selectedIndex, names);
+        int newIndex = EditorGUILayout.Popup(label, selectedIndex, names, options);
         
         if (selectedIndex != newIndex) {
             obj.enumValueIndex = newIndex;
@@ -731,9 +818,9 @@ public class MadGUI {
         runnable();
         GUI.enabled = savedState;
     }
-    
+
     public static void PropertyFieldObjectsPopup<T>(Object target, string label, ref T selectedObject, List<T> objects,
-        bool allowEditWhenDisabled) where T : Component {
+        bool allowEditWhenDisabled) where T : UnityEngine.Object {
         
         bool active = allowEditWhenDisabled || 
 #if UNITY_3_5
@@ -802,6 +889,8 @@ public class MadGUI {
         EditorGUILayout.EndHorizontal();
     }
 #endregion
+
+    #region Delegates
     
     public delegate void RunnableVoid0();
     public delegate void RunnableVoid1<T>(T arg1);
@@ -810,6 +899,10 @@ public class MadGUI {
     public delegate T RunnableGeneric1<T, T1>(T1 arg1);
     public delegate bool Validator(SerializedProperty property);
     public delegate bool Validator0();
+
+    #endregion
+
+    #region Types
     
     public abstract class ScrollableListItem {
         public bool selected;
@@ -836,6 +929,18 @@ public class MadGUI {
     public static Validator ObjectIsSet = (SerializedProperty s) => {
         return s.objectReferenceValue != null;
     };
+
+    public class IndentC : System.IDisposable {
+        public IndentC() {
+            EditorGUI.indentLevel++;
+        }
+
+        public void Dispose() {
+            EditorGUI.indentLevel--;
+        }
+    }
+
+    #endregion
 }
 
 } // namespace
